+++ src/main/java/com/programmers/board/controller/UserController.java
@@ -7,7 +7,10 @@ import com.programmers.board.controller.response.Result;
 import com.programmers.board.dto.UserDto;
 import com.programmers.board.dto.request.UserCreateRequest;
 import com.programmers.board.dto.request.UserUpdateRequest;
-import com.programmers.board.dto.service.*;
+import com.programmers.board.dto.service.user.UserCreateCommand;
+import com.programmers.board.dto.service.user.UserDeleteCommand;
+import com.programmers.board.dto.service.user.UserGetCommand;
+import com.programmers.board.dto.service.user.UserUpdateCommand;
 import com.programmers.board.exception.AuthenticationException;
 import com.programmers.board.service.UserService;
 import jakarta.validation.Valid;
