+++ src/main/java/com/programmers/board/service/UserService.java
@@ -1,12 +1,12 @@
 package com.programmers.board.service;
 
 import com.programmers.board.constant.AuthErrorMessage;
-import com.programmers.board.dto.service.UserDeleteCommand;
-import com.programmers.board.dto.service.UserGetCommand;
+import com.programmers.board.dto.service.user.UserDeleteCommand;
+import com.programmers.board.dto.service.user.UserGetCommand;
 import com.programmers.board.domain.User;
 import com.programmers.board.dto.UserDto;
-import com.programmers.board.dto.service.UserCreateCommand;
-import com.programmers.board.dto.service.UserUpdateCommand;
+import com.programmers.board.dto.service.user.UserCreateCommand;
+import com.programmers.board.dto.service.user.UserUpdateCommand;
 import com.programmers.board.exception.AuthorizationException;
 import com.programmers.board.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
